@@ -27,10 +27,11 @@
     "intl": "^1.2.5",
     "lodash": "^4.16.6",
     "normalize.css": "^5.0.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-fastclick": "^3.0.1",
-    "react-redux": "^4.4.5",
+    "react-redux": "^5.0.6",
     "react-router": "^3.0.0",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
